+++ src/time/period.rs
@@ -2,5 +2,5 @@ use super::timeunit::TimeUnit;
 
 pub struct Period {
     pub units: TimeUnit,
-    pub length: i32,
+    pub length: usize,
 }
